pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_marketplace"
version = "0.1.0"
description = "A two-sided marketplace platform for user's personal AI and service agents"
readme = "README.md"
authors = [{name = "PIN AI Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pydantic>=1.8.2",
    "python-dotenv>=0.19.0",
    "pydantic-settings>=2.8.1",
    "streamlit==1.43.0",
    "openai==1.65.4",
]

[project.optional-dependencies]
dev = ["pytest>=6.0", "black>=21.5b2", "isort>=5.9.1", "mypy>=0.812"]

[tool.setuptools]
packages = ["agent_marketplace"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
disallow_incomplete_defs = true 